requirements.txt
# Core dependencies
crewai>=0.1.0
langchain-openai>=0.1.0
langchain>=0.1.0
openai>=1.0.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Korean language processing
konlpy>=0.6.0
hangul-toolkit>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.0.0